Near miss study
A basic search idea used on a commissioned django site project. Originally was intended to be its own thing but ended up as a decent idea due to the site's need for mobile friendly features and setups. In my personal experience, this would capture many of the typos I make in my daily life with my phone.

The idea originated from reading descriptions of warship combat in WW1/WW2 which helped inspire the name and function. In testing with the project it is a part of, it has proven fairly reliable, though it is only compliant with english keyboards/input currently